Add files via upload
a.ss <-- a new version of a.s that creates a very small hash from shasum. The hash is in a shifted octal format, which I first called "Shoctal". 

a.fss <-- The analog of a.fs, this writes a given function into a named text file with the above mentioned extension. 

This starts to move toward "fiona", which has mirrored functionality in javascript. The "sum" and "cksum" commands are ancient unix tools. This curiously useful small (collision possible) hash seems well suited for blockchain data via the use of (advanced) satoshi codes. By this, I mean to say, five to seven decimal digits. Four is usually too small (and should be padded). The rightmost digits should be "02" for various reasons.